requirements.txt
astroid==2.1.0
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.5
django-allauth==0.38.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.0
gunicorn==19.9.0
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.1.0
Pillow==5.4.0
pylint==2.2.2
pylint-django==2.0.5
pylint-plugin-utils==0.4
python3-openid==3.1.0
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.0.0
six==1.12.0
urllib3==1.24.1
whitenoise==4.1.2
wrapt==1.10.11
psycopg2==2.7.2